Working on fixing warnings

+++ replay-agent/src/sdr.cc
@@ -19,31 +19,26 @@
  *
  */
 
-#include <spdlog/spdlog.h>
 #include "sdr.h"
 #include "exceptions.h"
+#include <spdlog/spdlog.h>
 
-/** 
+/**
  * Constructor for sdr.
  *
- * @param sample_rate 
- * @param frequency 
- * @param rx_gain 
- * @param tx_gain 
+ * @param sample_rate
+ * @param frequency
+ * @param rx_gain
+ * @param tx_gain
  */
-sdr::sdr(
-  double sample_rate,
-  double frequency,
-  string rf_args,
-  double rx_gain,
-  double tx_gain
-) : sample_rate(sample_rate),
-    frequency(frequency),
-    rx_gain(rx_gain),
-    tx_gain(tx_gain) {
+sdr::sdr(double sample_rate, double frequency, string rf_args, double rx_gain,
+         double tx_gain)
+    : sample_rate(sample_rate), frequency(frequency), rx_gain(rx_gain),
+      tx_gain(tx_gain) {
 
   // Try to open device
-  if (srsran_rf_open_devname(&rf, "", (char*)rf_args.c_str(), 1) == SRSRAN_ERROR) {
+  if (srsran_rf_open_devname(&rf, "", (char *)rf_args.c_str(), 1) ==
+      SRSRAN_ERROR) {
     throw sdr_exception("Failed to open SDR");
   }
 
@@ -57,10 +52,11 @@ sdr::sdr(
   secs_prev = 0;
   frac_secs_prev = 0;
 
-  SPDLOG_INFO("Started SDR on {} frequency, at {} sps, RX gain {}.", set_freq, set_sr, get_rx_gain);
+  SPDLOG_INFO("Started SDR on {} frequency, at {} sps, RX gain {}.", set_freq,
+              set_sr, get_rx_gain);
 }
 
-/** 
+/**
  * Destructor for sdr.
  */
 sdr::~sdr() {
@@ -68,35 +64,43 @@ sdr::~sdr() {
   srsran_rf_close(&rf);
 }
 
-/** 
+/**
  * Receive a vector of samples from the opened SDR.
  *
  * @param num_samples number of samples to receive
  */
 shared_ptr<vector<complex<float>>> sdr::receive(size_t num_samples) {
-  SPDLOG_DEBUG("RX {0} samples ({1:.3f} ms)", num_samples, num_samples / this->sample_rate * 1000.0);
+  SPDLOG_DEBUG("RX {0} samples ({1:.3f} ms)", num_samples,
+               num_samples / this->sample_rate * 1000.0);
 
   // Create a shared pointer holding the buffer of num_samples samples
-  shared_ptr<vector<complex<float>>> p = make_shared<vector<complex<float>>>(vector<complex<float>>(num_samples));
+  shared_ptr<vector<complex<float>>> p =
+      make_shared<vector<complex<float>>>(vector<complex<float>>(num_samples));
   time_t secs;
   double frac_secs;
 
   try {
-    int num_received_samples = srsran_rf_recv_with_time(&rf, p.get()->data(), num_samples, 1, &secs, &frac_secs);
-    SPDLOG_DEBUG("RF recv {} samples time secs {}, {}, diff with prev {},{}", num_samples, secs,frac_secs, secs - secs_prev, frac_secs - frac_secs_prev);
+    size_t num_received_samples = srsran_rf_recv_with_time(
+        &rf, p.get()->data(), num_samples, 1, &secs, &frac_secs);
+    SPDLOG_DEBUG("RF recv {} samples time secs {}, {}, diff with prev {},{}",
+                 num_samples, secs, frac_secs, secs - secs_prev,
+                 frac_secs - frac_secs_prev);
     secs_prev = secs;
     frac_secs_prev = frac_secs;
-    // TODO Bug in srsRAN?: even successful trials are counted as erroneous ones. So if we reach 100 trials the function returns -1 and stops receiving even though the data looks good.
+    // TODO Bug in srsRAN?: even successful trials are counted as erroneous
+    // ones. So if we reach 100 trials the function returns -1 and stops
+    // receiving even though the data looks good.
     if (num_received_samples != num_samples) {
-      if (num_received_samples == -1) {
+      if (num_received_samples == (size_t)-1) {
         throw sdr_exception("Error receiving samples");
       } else {
         stringstream ss;
-        ss << "Requested " << num_samples << " samples but only got " << num_received_samples;
+        ss << "Requested " << num_samples << " samples but only got "
+           << num_received_samples;
         throw sdr_exception(ss.str());
       }
     }
-  } catch(sdr_exception& e) {
+  } catch (sdr_exception &e) {
     SPDLOG_ERROR(e.what());
   }
